requirements.txt
numpy==1.19.0
opencv-python==4.5.0
tensorflow==2.4.0
fastapi==0.115.12
uvicorn==0.33.0
python-multipart==0.0.20
Pillow==10.4.0
pydantic==2.10.6
h5py==2.10.0
matplotlib==3.6.3
scikit-image==0.16.2
imgaug==0.4.0
